+++ src/core/client/systems/toolbar.ts
@@ -15,6 +15,10 @@ export class ToolbarController {
             return;
         }
 
+        if (alt.Player.local.isPhoneOpen) {
+            return;
+        }
+
         const slot: number = parseInt(String.fromCharCode(key)) - 1;
         const item = alt.Player.local.meta.toolbar.find((item) => item.slot === slot);