+++ documentation/changelog.md
@@ -12,6 +12,7 @@ description: List of all the changes and current versions.
 + Added Timeout to Finish Authorization Button (3 Seconds)
 + Fixed Fuel Cash Decimal Overflow
 + Disable Attacks while Phone is in Use
++ Disable Toolbar while Phone is in Use
 ```
 
 ## 1.8.0 - February 28-x, 2021